20260305 @Mookse wip - backend pipeline for direct access - frontend wip
