+++ inc/js/controllers/functions.mjs
@@ -19,21 +19,6 @@ async function about(ctx){
 	}
 }
 /**
- * Activate a specific Bot.
- * @public
- * @async
- * @param {object} ctx - Koa Context object
- * @returns {object} - Activated Response object: { bot_id, greeting, success, version, versionUpdate, }
- */
-async function activateBot(ctx){
-	const { bid, } = ctx.params
-	if(!ctx.Globals.isValidGuid(bid))
-		ctx.throw(400, `missing bot id`)
-	const { avatar: Avatar, } = ctx.state
-	const response =await Avatar.setActiveBot(bid)
-	ctx.body = response
-}
-/**
  * Get alerts (or specific alert) for the member/visitor.
  * @param {Koa} ctx - Koa Context object
  * @returns {Object[]} - The array of alerts
@@ -45,49 +30,6 @@ async function alerts(ctx){
 		ctx.body = await Avatar.alert(aid)
 	else
 		ctx.body = await Avatar.alerts()
-}
-/**
- * Manage bots for the member.
- * @param {Koa} ctx - Koa Context object
- * @returns {object} - Koa Context object
- */
-async function bots(ctx){
-	const { bid, } = ctx.params
-	const { avatar: Avatar, } = ctx.state
-	const bot = ctx.request.body
-		?? {}
-	switch(ctx.method){
-		case 'DELETE': // retire bot
-			if(!ctx.Globals.isValidGuid(bid))
-				ctx.throw(400, `missing bot id`)
-			ctx.body = await Avatar.retireBot(bid)
-			break
-		case 'POST': // create new bot
-			ctx.body = await Avatar.createBot(bot)
-			break
-		case 'PUT': // update bot
-			ctx.body = await Avatar.updateBot(bot)
-			break
-		case 'GET':
-		default:
-			if(bid?.length){ // specific bot
-				ctx.body = await Avatar.getBot(bid)
-			} else {
-				const bots = await Avatar.getBots()
-				let { activeBotId, greeting, } = Avatar
-				if(!activeBotId){
-					const { bot_id, greeting: activeGreeting } = await Avatar.setActiveBot()
-					activeBotId = bot_id
-					greeting = activeGreeting
-				}
-				ctx.body = { // wrap bots
-					activeBotId,
-					bots,
-					greeting,
-				}
-			}
-			break
-	}
 }
 /**
  * Challenge the member session with a passphrase.
@@ -119,38 +61,10 @@ async function challenge(ctx, memberId, memberPassphrase){
 	}
 	ctx.body = !ctx.session.locked
 }
-/**
- * Chat with the Member or System Avatar's intelligence.
- * @public
- * @async
- * @param {Koa} ctx - Koa Context object
- * @returns {object} - The response from the chat in `ctx.body`
- * @property {object} instruction - Instructionset for the frontend to execute (optional)
- * @property {Object[]} responses - Response messages from Avatar intelligence
- */
-async function chat(ctx){
-	const { botId: bot_id, itemId, message, } = ctx.request.body
-		?? {} /* body nodes sent by fe */
-	if(!message?.length)
-			ctx.throw(400, 'missing `message` content')
-	const { avatar: Avatar, } = ctx.state
-	if(bot_id?.length && bot_id!==Avatar.activeBotId)
-		throw new Error(`Bot ${ bot_id } not currently active; chat() requires active bot`)
-	const response = await Avatar.chat(message, itemId, ctx.session)
-	ctx.body = response
-}
 async function collections(ctx){
 	const { type, } = ctx.params
 	const { avatar, } = ctx.state
 	ctx.body = await avatar.collections(type)
-}
-async function createBot(ctx){
-	const { teamId, type, } = ctx.request.body
-	const { avatar, } = ctx.state
-	const bot = { teams: [], type, } // `type` only requirement to create a known, MyLife-typed bot
-	if(teamId?.length)
-		bot.teams.push(teamId)
-	ctx.body = await avatar.createBot(bot)
 }
 /**
  * Given an itemId, evaluates aspects of contents of the data record.
@@ -220,7 +134,6 @@ async function index(ctx){
 async function item(ctx){
 	const { iid: id, } = ctx.params
 	const { avatar, } = ctx.state
-	const { globals, } = avatar
 	const { method, } = ctx.request
 	const item = ctx.request.body // always `{}` by default
 	if(!item?.id && id?.length)
@@ -252,16 +165,6 @@ async function loginSelect(ctx){
 }
 async function members(ctx){ // members home
 	await ctx.render('members')
-}
-async function migrateBot(ctx){
-	const { bid, } = ctx.params
-	const { avatar, } = ctx.state
-	ctx.body = await avatar.migrateBot(bid)
-}
-async function migrateChat(ctx){
-	const { bid, } = ctx.params
-	const { avatar, } = ctx.state
-	ctx.body = await avatar.migrateChat(bid)
 }
 /**
  * Given an itemId, obscures aspects of contents of the data record.
@@ -300,45 +203,6 @@ async function privacyPolicy(ctx){
 		const response = await Avatar.routine('privacy')
 		ctx.body = response
 	}
-}
-/**
- * Direct request from member to retire a bot.
- * @param {Koa} ctx - Koa Context object
- */
-async function retireBot(ctx){
-	ctx.method = 'DELETE'
-	return await this.bots(ctx)
-}
-/**
- * Direct request from member to retire a chat (via bot).
- * @param {Koa} ctx - Koa Context object
- */
-async function retireChat(ctx){
-	const { avatar: Avatar, } = ctx.state
-	const { bid, } = ctx.params
-	if(!bid?.length)
-		ctx.throw(400, `missing bot id`)
-	const response = await Avatar.retireChat(bid)
-	ctx.body = response
-}
-/**
- * Routines are pre-composed scripts that can be run on-demand. They animate HTML content formatted by <section>.
- * @param {Koa} ctx - Koa Context object
- */
-async function routine(ctx){
-	const { rid, } = ctx.params
-	const { avatar: Avatar, } = ctx.state
-	const response = await Avatar.routine(rid)
-	ctx.body = response
-}
-/**
- * Gets the list of shadows.
- * @returns {Object[]} - Array of shadow objects.
- */
-async function shadows(ctx){
-	const { avatar, } = ctx.state
-	const response = await avatar.shadows()
-	ctx.body = response
 }
 async function signup(ctx) {
     const { avatarName, email, humanName, type='newsletter', } = ctx.request.body
@@ -412,17 +276,6 @@ async function teams(ctx){
 	const { avatar: Avatar, } = ctx.state
 	ctx.body = await Avatar.teams()
 }
-async function updateBotInstructions(ctx){
-	const { bid, } = ctx.params
-	if(!bid?.length)
-		ctx.throw(400, `missing bot id`)
-	const { avatar, } = ctx.state
-	const bot = await avatar.updateBotInstructions(bid)
-	ctx.body = {
-		bot,
-		success: !!bot,
-	}
-}
 /**
  * Proxy for uploading files to the API.
  * @param {Koa} ctx - Koa Context object
@@ -439,13 +292,9 @@ async function upload(ctx){
 /* exports */
 export {
 	about,
-	activateBot,
 	alerts,
-	bots,
 	challenge,
-	chat,
 	collections,
-	createBot,
 	evaluate,
 	feedback,
 	greetings,
@@ -455,19 +304,12 @@ export {
 	logout,
 	loginSelect,
 	members,
-    migrateBot,
-    migrateChat,
 	obscure,
 	passphraseReset,
 	privacyPolicy,
-	retireBot,
-	retireChat,
-	routine,
-	shadows,
 	signup,
 	summarize,
 	team,
 	teams,
-	updateBotInstructions,
 	upload,
 }
