+++ inc/json-schemas/intelligences/biographer-intelligence-1.9.json
@@ -190,6 +190,9 @@
           "value": "travel"
         }
       ],
+      "order": 0,
+      "placeholder": null,
+      "range": {},
       "title": "My Interests",
       "type": "checkbox",
       "variable": "interests"
